views.py is divided in 3
