+++ canteen_menu/menu/templates/search.html
@@ -43,7 +43,7 @@
                                          <span class="range-selected"></span>
                                        </div>
                                        <div class="range-input">
-                                         <input type="range" name="cal-min" class="min" min="0" max="1000" value="300" step="10">
+                                         <input type="range" name="cal-min" class="min" min="0" max="1000"   value="300" step="10">
                                          <input type="range" name="cal-max" class="max" min="0" max="1000" value="700" step="10">
                                        </div>
                                       </div>
