+++ canteen_menu/menu/templates/staff_master.html
@@ -2,7 +2,7 @@
 <html lang="en">
 <head>
   <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=2.0">
+  <meta name="viewport" content="width=device-width, initial-scale=1">
   <link rel="stylesheet" href="style.css">
   <style>
     @media (max-width: 767px) {
@@ -117,6 +117,8 @@
   <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js"></script>
   <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js"></script>
   <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js"></script>
+  <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.7.1/jquery.min.js"></script>
+  <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.11.2/font/bootstrap-icons.min.css">
 </head>
 
 <body>
@@ -136,38 +138,23 @@
       <div class="container-fluid text-light">
         <!-- Links -->
         <ul class="navbar-nav">
-          <li class="nav-item">
-            <a class="nav-brand font-weight-bold h1" style="text-decoration: none" href="/">LICĖJUS</a>
-          </li>
-          <li class="nav-item">
+          <li class="nav-item me-5">
+              <a class="nav-brand font-weight-bold h1" style="text-decoration: none" href="/">LICĖJUS</a>
           </li>
         </ul>
       </div>
 
       <div class="collapse navbar-collapse justify-content-end" id="navbarSupportedContent">
         <ul class="navbar-nav ml-auto">
-          {% if user.is_authenticated %}
-            {% if user.is_staff %}
-              <li>
-                <a class="nav-link active h1" href="/staff">Go to staff</a>
-              </li>
-            {% endif %}
-            <li class="nav-item dropdown">
-              <a class="nav-link dropdown-toggle active" role="button" data-bs-toggle="dropdown">Profile</a>
-                <ul class="dropdown-menu align-right dropdown-menu-end" style="right: 0">
-                  <li><p class="dropdown-item active font-weight-bold mb-0">Welcome, {{ user.username }}</p></li>
-                  <li><a class="dropdown-item" href="/favorite">Favorites</a></li>
-                  <li><a class="dropdown-item" href="/logout">Log out</a></li>
-                </ul>
-            </li>
-          {% else %}
-            <li class="nav-item">
-              <a class="nav-link active h1" style="text-decoration: none; font-size: 1.5rem;" href="/register">Register</a>
-            </li>
-            <li class="nav-item">
-              <a class="nav-link active h1" style="text-decoration: none; font-size: 1.5rem;" href="/login">Login</a>
-            </li>
-          {% endif %}
+          <li class="nav-item mt-2">
+              <a class="nav-link active h1" style="text-decoration: none" href="/staff">Meals</a>
+          </li>
+          <li class="nav-item mt-2">
+              <a class="nav-link active h1" style="text-decoration: none" href="/staff/schedule">Schedule</a>
+          </li>
+          <li>
+              <a class="nav-link active h1" href="/">Go to main</a>
+          </li>
         </ul>
       </div>
   </nav>
