+++ canteen_menu/menu/templates/staff_meals.html
@@ -10,6 +10,7 @@
     <!-- The user is authenticated and is an admin -->
     <div class="container mt-4">
         <div class="row d-flex align-items-center justify-content-center">
+<!--            <h1>Meals</h1>-->
             <table class="table">
                 <a type="button" class="btn btn-success" href="staff/meal/add/">Add</a>
                 <thead>
@@ -54,7 +55,7 @@
     </div>
     {% else %}
     <!-- The user is not an admin or is not authenticated -->
-    <p>Пашоу нахой</p>
+    <p>Page is blocked: you are not a staff</p>
     {% endif %}
     <script>
         $(document).ready(function () {
